Improved UI for CustomLists
