+++ Shared/View/Details/PersonDetailsView.swift
@@ -135,7 +135,7 @@ struct PersonDetailsView: View {
 #else
             Table(viewModel.credits.filter { ($0.itemTitle.localizedStandardContains(viewModel.query)) as Bool }) {
                 TableColumn("Title") { item in
-                    SearchItemView(item: item, showInformationPopup: $showInformationPopup)
+                    SearchItemView(item: item, showConfirmation: $showSaveConfirmation)
                         .buttonStyle(.plain)
                         .accessibilityHint(Text(item.itemTitle))
                 }
