+++ Shared/View/Navigation/HomeView.swift
@@ -82,9 +82,9 @@ struct HomeView: View {
             .navigationDestination(for: ProductionCompany.self) { item in
                 CompanyDetails(company: item)
             }
-            .navigationDestination(for: [ProductionCompany].self, destination: { item in
+            .navigationDestination(for: [ProductionCompany].self) { item in
                 CompaniesListView(companies: item)
-            })
+            }
             .redacted(reason: !viewModel.isLoaded ? .placeholder : [] )
             .navigationTitle("Home")
             .toolbar {
